MaintenanceRepositoryTests Query

+++ HW1/src/main/java/com/example/demo/repositories/CorrectiveRepairsRepository.java
@@ -1,14 +1,18 @@
 package com.example.demo.repositories;
 
+import com.example.demo.models.MaintenancePlan;
+import com.example.demo.models.PlantInventoryItem;
 import com.example.demo.utils.Pair;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CorrectiveRepairsRepository {
 
 	List<Pair<Integer, Long>> findCorrectiveRepairsByYearForPeriod(int startYear, int endYear);
 
-	List<Pair<Integer, Integer>> findCorrectiveRepairCostsByYear(int startYear, int endYear);
+	List<Pair<Integer, BigDecimal>> findCorrectiveRepairCostsByYear(int startYear, int endYear);
+
 
 
 }
